+++ NovaBridge/Source/NovaBridgeCore/NovaBridgeCore.Build.cs
@@ -10,6 +10,7 @@ public class NovaBridgeCore : ModuleRules
 		{
 			"Core",
 			"Json",
+			"HTTPServer",
 		});
 	}
 }